Remove redundant diff request and clean normalize header
